Replace WrapResult trait by returning Context object
So, there's an easier way to get the HTTP request, result, and other
context information to the user than returning a WrappedResult object.
Just return the Twitter API result in scalar context, and both it and
the context in an array context.

Make the context hash an object with appropriate accessors so it behaves
just like the WrappedResult object did.