Use MooX::Traits
This results in a bunch of ugly code to handle psuedo trait
'@enchilada'. It was cute, but not worth having to wrap new_with_traits
when I can just create a trait the includes the others and let
MooX::Traits do it's thing.

This also means users have to use new_with_traits, rather than new.
Which is probably ok, but it results in a bigger change for anyone
switching from Net::Twitter. :-(