Fix invalidate_token
Dammit, Twitter! When you send me application/json, I expect the decoded
JSON to contain decoded values. But you lack trust. You know you'll want
the token back URL encoded. So, instead of sending a raw token in JSON,
you send a URL encoded token.

Then, to revoke it, you want an application/x-www-form-urlencoded POST.
A properly behaved application URL encodes its parameters. But if we
URL encode the token, it's now double-encoded and you choke on it.

So, URL decode the token Twitter gives us in JSON and properly encoded
it when we send it back.