Dead End
Making this an upgrade to Net::Twitter seems fraught with disaster. So,
at least for the moment, I'm abandoning this approach.

My plan was to write a shim in Net::Twitter v4.x that warns about
incompatible v5.x construction and accepts v5.x new args, transforming
them to v4.x equivalents so the v4.x users could update their code
for v5.x compatibility before the v5.x release.

For v5.x, write a shim that accepts v4.x "new" args, warns, and
transforms if possible mitigating the surprise of users when upgrade
unaware of the interface incompatibility.

That's a mountain of work. Without it, an upgrade will potentially break
a lot of code and make people very unhappy.

I hate hove THREE modules out there that all solve the same problem, but
it seems the least work for me and the least surprise and disruption for
anyone making the switch.

Also:
- fixed a couple of renames I missed

[skip ci]