Partial implementation of Twitter API V2 support
Experimental. This will change!

The Twitter Open API spec at https://api.twitter.com/2/openapi.json is
not, at this time, stable and consistent. It may be possible to
automatically generate client code from the spec at some point in the
future. But at this time, the spec is a rough guide. I've hand coded
support for a portion. It's a first, incomplete draft.