It's not 200 OK, Twitter!
Twitter API v2 is returning 200 OK in some error conditions. So, when
Twitter gives us a 200 OK with errors in the response body but no data,
modify the response and throw an exception.

Use a best guess to translate Twitter's problem type to a valid HTTP
status code. Add a custom header to the response indicating Twitter's
original status.

Hopefully, I can remove this code if/when Twitter returns a valid HTTP
status code for errors.